// src/models/user.model.ts
import mongoose, { Schema, Document } from 'mongoose';
import type { User } from '../interfaces/user.interface';
import { hashPassword } from '../utils/bcrypt';

export interface UserDocument extends Omit<User, '_id'>, Document {}

const userSchema = new Schema<UserDocument>(
  {
    email: {
      type: String,
      required: true,
      unique: true,
      lowercase: true,
      match: /^[^\s@]+@[^\s@]+\.[^\s@]+$/,
    },
    password: {
      type: String,
      required: true,
      minlength: 6,
    },
    firstName: {
      type: String,
      required: true,
    },
    lastName: {
      type: String,
      required: true,
    },
  },
  {
    timestamps: true,
  }
);

// Hash password before saving if it has been modified
userSchema.pre('save', async function (next) {
  // Only hash if password is new or has been modified
  if (!this.isModified('password')) {
    return next();
  }

  try {
    this.password = await hashPassword(this.password);
    next();
  } catch (error) {
    next(error as Error);
  }
});

export const UserModel = mongoose.model<UserDocument>('User', userSchema);
